track report finalization state in controller.
since report.isFinalized() is deprecated and going away without
replacement, we'll need to track that state at the highest level here,
in the controllers.
this requires us to push functionality from the "top" ci report route
and controller into the sub-routes/controllers. which leads to
duplication of functionality, but there's no way around that.
