+++ app/routes/curriculum-inventory-report/index.js
@@ -1,11 +1,9 @@
-import { inject as service } from '@ember/service';
 import Route from '@ember/routing/route';
+import { inject as service } from '@ember/service';
 
-export default class CurriculumInventoryReportIndexRoute extends Route {
-  @service store;
+export default class CurriculumInventoryReportReport extends Route {
   @service permissionChecker;
   @service session;
-
   canUpdate = false;
 
   beforeModel(transition) {
@@ -13,16 +11,14 @@ export default class CurriculumInventoryReportIndexRoute extends Route {
   }
 
   async afterModel(report) {
-    await report.get('topLevelSequenceBlocks');
-
     const permissionChecker = this.permissionChecker;
     const canUpdate = await permissionChecker.canUpdateCurriculumInventoryReport(report);
-
     this.set('canUpdate', canUpdate);
   }
 
   setupController(controller, model) {
     super.setupController(controller, model);
-    controller.set('canUpdate', this.canUpdate);
+    controller.set('hasUpdatePermissions', this.canUpdate);
+    controller.set('isFinalized', !!model.belongsTo('export').id());
   }
 }
